refactor: move to new folder structure
