+++ templates/vue/src/components/Sidebar/NodeReview/ReviewButtons.vue
